requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
numpy>=1.24.0
matplotlib==3.8.2
gunicorn==21.2.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
pytz==2024.1
dnspython==2.4.2
APScheduler==3.10.4
scholarly==1.7.11
habanero==1.2.6
